refactor: restructure project layout
