+++ app/poll/context.go
@@ -1,4 +1,4 @@
-package store
+package poll
 
 import "context"
